--- br/pkg/lightning/backend/kv/BUILD.bazel
@@ -1,78 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "kv",
-    srcs = [
-        "allocator.go",
-        "kv2sql.go",
-        "session.go",
-        "sql2kv.go",
-        "types.go",
-    ],
-    importpath = "github.com/pingcap/tidb/br/pkg/lightning/backend/kv",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//br/pkg/lightning/common",
-        "//br/pkg/lightning/log",
-        "//br/pkg/lightning/manual",
-        "//br/pkg/lightning/metric",
-        "//br/pkg/lightning/verification",
-        "//br/pkg/logutil",
-        "//br/pkg/redact",
-        "//br/pkg/utils",
-        "//expression",
-        "//kv",
-        "//meta/autoid",
-        "//parser/model",
-        "//parser/mysql",
-        "//sessionctx",
-        "//sessionctx/variable",
-        "//table",
-        "//table/tables",
-        "//tablecodec",
-        "//types",
-        "//util/chunk",
-        "//util/mathutil",
-        "//util/topsql/stmtstats",
-        "@com_github_docker_go_units//:go-units",
-        "@com_github_pingcap_errors//:errors",
-        "@org_golang_x_exp//slices",
-        "@org_uber_go_zap//:zap",
-        "@org_uber_go_zap//zapcore",
-    ],
-)
-
-go_test(
-    name = "kv_test",
-    timeout = "short",
-    srcs = [
-        "session_internal_test.go",
-        "session_test.go",
-        "sql2kv_test.go",
-    ],
-    embed = [":kv"],
-    flaky = True,
-    race = "on",
-    deps = [
-        "//br/pkg/lightning/common",
-        "//br/pkg/lightning/log",
-        "//br/pkg/lightning/verification",
-        "//ddl",
-        "//kv",
-        "//meta/autoid",
-        "//parser",
-        "//parser/ast",
-        "//parser/model",
-        "//parser/mysql",
-        "//sessionctx",
-        "//table",
-        "//table/tables",
-        "//tablecodec",
-        "//types",
-        "//util/mock",
-        "@com_github_docker_go_units//:go-units",
-        "@com_github_stretchr_testify//require",
-        "@org_uber_go_zap//:zap",
-        "@org_uber_go_zap//zapcore",
-    ],
-)
